requirements.txt
pip==22.2.1
requests[socks]~=2.28.1
fake-useragent~=0.1.11
python-dotenv~=0.2
pika~=1.3.0